+++ src/annotations.ts
@@ -36,10 +36,11 @@ Portions of this software are derived from [ghcid](https://github.com/ndmitchell
 import { dirname } from 'path'
 import * as vscode from 'vscode'
 
-import { Annotation, AnnotationsConfig, HierarchicalOutputChannel, LanguageConfig, alloglot } from "./config";
+import { Annotation, AnnotationsConfig, LanguageConfig, alloglot } from './config'
+import { IHierarchicalOutputChannel } from './utils'
 
-export function makeAnnotations(output: HierarchicalOutputChannel, config: LanguageConfig): vscode.Disposable {
-  output.appendLine('Starting annotations...')
+export function makeAnnotations(output: IHierarchicalOutputChannel, config: LanguageConfig): vscode.Disposable {
+  output.appendLine(alloglot.ui.startingAnnotations)
   const { languageId, annotations } = config
   if (!languageId || !annotations || annotations.length === 0) return vscode.Disposable.from()
 
@@ -51,7 +52,7 @@ export function makeAnnotations(output: HierarchicalOutputChannel, config: Langu
     { providedCodeActionKinds: [vscode.CodeActionKind.QuickFix] }
   )
 
-  output.appendLine('Annotation started.')
+  output.appendLine(alloglot.ui.annotationsStarted)
   return vscode.Disposable.from(
     quickFixes,
     ...watchers