+++ src/apisearch.ts
@@ -206,19 +206,20 @@ Portions of this software are derived from [vscode-goto-documentation](https://g
 
 import * as vscode from 'vscode'
 
-import { Config, HierarchicalOutputChannel, alloglot } from './config'
+import { Config, alloglot } from './config'
+import { IHierarchicalOutputChannel } from './utils'
 
-export function makeApiSearch(output: HierarchicalOutputChannel, config: Config): vscode.Disposable {
+export function makeApiSearch(output: IHierarchicalOutputChannel, config: Config): vscode.Disposable {
   const { languages } = config
-  if (languages.length === 0) return vscode.Disposable.from()
+  if (!languages || languages.length === 0) return vscode.Disposable.from()
 
   const langs: Map<string, string> = new Map()
   languages.forEach(lang => {
     lang.languageId && lang.apiSearchUrl && langs.set(lang.languageId, lang.apiSearchUrl)
   })
 
-  output.appendLine(`Creating API search command for languages...`)
-  config.languages.forEach(lang => output.appendLine(`\t${lang.languageId}`))
+  output.appendLine(alloglot.ui.creatingApiSearch)
+  config.languages?.forEach(lang => output.appendLine(`\t${lang.languageId}`))
 
   return vscode.commands.registerTextEditorCommand(
     alloglot.commands.apiSearch,