+++ src/formatter.ts
@@ -23,64 +23,45 @@ Portions of this software are derived from [vscode-custom-local-formatters](http
 > OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 > SOFTWARE.
 */
-import { exec } from 'child_process'
 import * as vscode from 'vscode'
 
-import { HierarchicalOutputChannel, LanguageConfig, alloglot } from "./config";
+import { LanguageConfig, alloglot } from './config'
+import { AsyncProcess, Disposal, IHierarchicalOutputChannel } from './utils'
 
 /**
  * Register a custom document formatter for a language.
  */
-export function makeFormatter(output: HierarchicalOutputChannel, config: LanguageConfig): vscode.Disposable {
+export function makeFormatter(output: IHierarchicalOutputChannel, config: LanguageConfig): vscode.Disposable {
   const { languageId, formatCommand } = config
   if (!languageId || !formatCommand) return vscode.Disposable.from()
 
-  output.appendLine('Starting formatter...')
+  output.appendLine(alloglot.ui.startingFormatter)
+
+  const disposal = Disposal.make()
+
   const formatter = vscode.languages.registerDocumentFormattingEditProvider(
     languageId,
     {
       provideDocumentFormattingEdits: document => {
         const command = formatCommand.replace('${file}', document.fileName)
-        const cwd = utils.getWorkspaceFolder(document)
-        const documentText = document.getText()
+        const basedir = vscode.workspace.workspaceFolders?.[0].uri
+        const stdin = document.getText()
         const entireDocument = new vscode.Range(
           document.lineAt(0).range.start,
           document.lineAt(document.lineCount - 1).rangeIncludingLineBreak.end,
         );
 
-        output.append(`Formatting document with:\n\tcommand=${command}\n\tcwd=${cwd}\n`)
-
-        return new Promise<Array<vscode.TextEdit>>((resolve, reject) => {
-          const proc = exec(command, { cwd }, (error, stdout, stderr) => {
-            if (error) {
-              output.append(`Error running formatter:\t${error}\n`)
-              reject(error)
-            }
-            else if (!stdout) {
-              output.append(`Formatter produced no output.\n`)
-              resolve([])
-            }
-            else {
-              stderr && output.append(`Formatter logs:\n${stderr}\n`)
-              resolve([new vscode.TextEdit(entireDocument, stdout)])
-            }
-          })
-
-          proc.stdin?.write(documentText)
-          proc.stdin?.end()
-        })
+        const proc = AsyncProcess.make({ output, command, basedir, stdin }, stdout => [new vscode.TextEdit(entireDocument, stdout)])
+        disposal.insert(proc)
+        return proc
       }
     }
   )
 
-  output.appendLine('Formatter started.')
-  return formatter
-}
+  output.appendLine(alloglot.ui.formatterStarted)
 
-namespace utils {
-  export function getWorkspaceFolder(document: vscode.TextDocument): string | undefined {
-    const workspaceFolder = vscode.workspace.getWorkspaceFolder(document.uri)
-    const backupFolder = vscode.workspace.workspaceFolders?.[0]
-    return workspaceFolder?.uri?.fsPath || backupFolder?.uri.fsPath
-  }
+  return vscode.Disposable.from(
+    formatter,
+    disposal
+  )
 }